Mejoras-API

+++ flagsentinel-api/flagsentinel-api/src/main/java/com/example/flagsentinelapi/service/UserService.java
@@ -1,13 +1,18 @@
 package com.example.flagsentinelapi.service;
 
 import com.example.flagsentinelapi.dto.CreateUserRequest;
+import com.example.flagsentinelapi.dto.RuleResponse;
 import com.example.flagsentinelapi.dto.UpdateUserRequest;
 import com.example.flagsentinelapi.dto.UserResponse;
+import com.example.flagsentinelapi.exception.BadRequestException;
 import com.example.flagsentinelapi.exception.ConflictException;
 import com.example.flagsentinelapi.exception.NotFoundException;
 import com.example.flagsentinelapi.mapper.UserMapper;
 import com.example.flagsentinelapi.model.User;
 import com.example.flagsentinelapi.repository.UserRepository;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
@@ -65,9 +70,29 @@ public class UserService {
 
     public void delete(Long id) {
 
+        // 1. Obtener el username del usuario autenticado
+        String currentUsername = SecurityContextHolder.getContext().getAuthentication().getName();
+
+        // 2. Obtener el usuario autenticado desde la BD
+        User currentUser = repo.findByUsername(currentUsername)
+                .orElseThrow(() -> new NotFoundException("Authenticated user not found"));
+
+        // 3. Impedir que se borre a sí mismo
+        if (currentUser.getId().equals(id)) {
+            throw new BadRequestException("No puedes borrar tu propio usuario");
+        }
+
+        // 4. Borrar el usuario objetivo
         User user = repo.findById(id)
                 .orElseThrow(() -> new NotFoundException("User not found"));
 
         repo.delete(user);
     }
+
+    public Page<UserResponse> findAllPaged(Pageable pageable) {
+        return repo.findAll(pageable)
+                .map(mapper::toResponse);
+    }
+
+
 }
